separated types from configuration files
